Updates after testing with wxWidgets 2.9.3 under Linux Mint 12.
git-svn-id: svn://svn.code.sf.net/p/xlights/code/trunk@83 ad4834c4-2333-46e5-bf01-adba983f9b38

+++ xConverter/xConverter.h
@@ -85,7 +85,7 @@ class xConverter: public wxFrame
         wxUint8 * SeqData;
         long SeqDataLen;
         long SeqNumPeriods;
-        size_t SeqNumChannels;
+        long SeqNumChannels;
         size_t TotChannels;
         wxArrayString FileNames;
         wxArrayString ChannelNames;
